helper function pindah ke folder latihan, tambah main
