--- hello/hello.go
@@ -1,11 +1,0 @@
-package hello
-
-import "fmt"
-
-// 1. emmbuat file lain harus di simpan di dalam package / folder
-// 2. kita harus menamai package kita sesuai nama foldernya
-// 3. function harus diawali dengan huruf besar (wajibb)
-
-func Hello() {
-	fmt.Println("hello world")
-}
