+++ latihan/latihan-1.go
@@ -1,0 +1,7 @@
+package latihan
+
+func Belanja(total int) (ppn int, beforePpn int) {
+	beforePpn = total * 10 / 11
+	ppn = total / 11
+	return
+}
